src/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
fastapi==0.119.0
idna==3.11
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.5
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
